complete order,order details,cart,cart details,payment

+++ src/main/java/com/ffood/g1/entity/Cart.java
@@ -41,7 +41,7 @@ public class Cart {
     private String status;
 
     @Column(name = "total_amount", nullable = false)
-    private Double totalAmount;
+    private Integer totalAmount;
 
     @Override
     public String toString() {
