+++ src/main/java/com/ffood/g1/service/CartService.java
@@ -3,7 +3,6 @@ package com.ffood.g1.service;
 
 
 import com.ffood.g1.entity.Cart;
-import com.ffood.g1.entity.Food;
 import com.ffood.g1.entity.User;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,7 +11,7 @@ import java.time.LocalDateTime;
 public interface CartService {
 
     Cart getOrCreateCart(User user);
-    void addToCart(Cart cart, Integer foodId, int quantity, LocalDateTime transactionDate, Double price);
+    void addToCart(Cart cart, Integer foodId, int quantity, LocalDateTime transactionDate, Integer price);
     Cart getCartByUserId(Integer userId); // Phương thức này phải được triển khai
 
     void removeCartItem(Integer cartItemId);
@@ -22,7 +21,7 @@ public interface CartService {
     @Transactional
     Integer findCartIdByUserId(Integer userId);
 
-    double getTotalFoodPriceByCartId(Integer cartId);
+    Integer getTotalFoodPriceByCartId(Integer cartId);
 
     @Transactional
     void clearCart(Cart cart);
